Initial implementation of LBDataMatrix
Some Swift syntax changes enforced by XCode 6.3.1
Added classes missing from LabBot.h
DRHLabJack.h and DRHLabJackU6 removed from LabBot.h because of "include of non-modular header" compile error that relates to labjackusb.h.  Not fixed by allow non-modular compiler flag.

+++ DRHGridView.swift
@@ -14,7 +14,7 @@ extension String {
         let lastChar = self.substringFromIndex(self.endIndex.predecessor())
         var remString = self.substringToIndex(self.endIndex.predecessor())
         let lastCharUnicode = lastChar.unicodeScalars[lastChar.unicodeScalars.startIndex].value
-        let newUnicode = lastCharUnicode + increment
+        let newUnicode = lastCharUnicode + UInt32(increment)
         var newLastChar = Character(UnicodeScalar(newUnicode))
         let ceilingUnicode = String(ceiling).unicodeScalars[lastChar.unicodeScalars.startIndex].value
         if newUnicode > ceilingUnicode {
